ParserTest.javaf
Prototype of junit testing of parser.
We need three kinds of comparison.
Simplest is to create a print routine and compare the output.
We may need to do a recursive comparison by overriding equals.
We may need to do the same, with variables replacement. Every consistent renaming is OK, as long as it does not collapse two variables into one.